+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ﻿# DocuSign C# Client Changelog
 
+## [v8.0.0-rc2] - eSignature API v2.1-24.2.00.00 - 2024-07-23
+### Changed
+- Introduced async versions of `ApiClient` authorization methods.
+- Corrected SDK metadata.
+- Updated the SDK release version.
 ## [v8.0.0-rc1] - eSignature API v2.1-24.2.00.00 - 2024-07-02
 
 ### Breaking Changes